api/users/serializers.py
from djoser.serializers import UserSerializer as BaseUserSerializer, UserCreateSerializer as BaseUserCreateSerializer

class UserCreateSerializer(BaseUserCreateSerializer):
    pass

class UserSerializer(BaseUserSerializer):
    pass