api/requirements_dev.txt
-r requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
colorama==0.4.4
coverage==6.3.2
docopt==0.6.2
iniconfig==1.1.1
model-bakery==1.5.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-watch==4.2.0
toml==0.10.2
tomli==2.0.1
watchdog==2.1.7